+++ src/lobpcg.cu
@@ -47,7 +47,7 @@ void lobpcg(
     double* D,       // m x m, device pointer (output eigenvalues, diagonal)
     const int n,
     const int m ,      // number of eigenpairs
-    const int warmstart,
+    const double warmstart,
     const int maxiter, // maximum iterations
     const double tol,   // convergence tolerance
     const bool verbose
